+++ packages/client/ui/react-ui/src/components/auth/methods/google/ContinueWithGoogle.tsx
@@ -1,12 +1,12 @@
 import type { UIConfig } from "@crossmint/common-sdk-base";
 import { Spinner } from "@/components/common/Spinner";
-import { useOAuthWindowListener } from "@/hooks/useOAuthWindowListener";
+import { useOAuthFlow } from "@/providers/auth/OAuthFlowProvider";
 import { GoogleIcon } from "@/icons/google";
 import { classNames } from "@/utils/classNames";
 import { tw } from "@/twind-instance";
 
 export function ContinueWithGoogle({ emailInput, appearance }: { emailInput: string; appearance?: UIConfig }) {
-    const { createPopupAndSetupListeners, isLoading: isLoadingOAuthWindow } = useOAuthWindowListener("google");
+    const { startOAuthLogin, isLoading } = useOAuthFlow();
 
     return (
         <button
@@ -14,18 +14,16 @@ export function ContinueWithGoogle({ emailInput, appearance }: { emailInput: str
             className={classNames(
                 "flex items-center gap-2 justify-center h-[32px] px-2.5 border border-cm-border rounded-xl bg-cm-background-primary",
                 "hover:bg-cm-hover focus:bg-cm-hover outline-none",
-                isLoadingOAuthWindow ? "cursor-not-allowed hover:bg-cm-muted-primary" : ""
+                isLoading ? "cursor-not-allowed hover:bg-cm-muted-primary" : ""
             )}
-            onClick={
-                isLoadingOAuthWindow ? undefined : () => createPopupAndSetupListeners(emailInput.trim().toLowerCase())
-            }
+            onClick={isLoading ? undefined : () => startOAuthLogin("google", emailInput.trim().toLowerCase())}
             style={{
                 backgroundColor: appearance?.colors?.buttonBackground,
                 borderRadius: appearance?.borderRadius,
             }}
         >
             <GoogleIcon className={tw("max-h-[18px] max-w-[18px] h-[18px] w-[18px]")} />
-            {isLoadingOAuthWindow ? (
+            {isLoading ? (
                 <Spinner
                     style={{
                         color: appearance?.colors?.textSecondary,