+++ packages/client/ui/react-ui/src/providers/auth/AuthFormProvider.tsx
@@ -1,17 +1,9 @@
-import { createContext, useCallback, useContext, useEffect, useState, type ReactNode } from "react";
-import { CrossmintAuthenticationError, type OAuthProvider } from "@crossmint/common-sdk-auth";
+import { createContext, useContext, useState, type ReactNode } from "react";
 import type { UIConfig } from "@crossmint/common-sdk-base";
-import { useCrossmintAuth } from "@/hooks/useCrossmintAuth";
 import type { BaseCrossmintWalletProviderProps, LoginMethod } from "@/types/auth";
 
 type AuthStep = "initial" | "otp" | "qrCode" | "web3" | "web3/metamask" | "web3/walletconnect";
 
-type OAuthUrlMap = Record<OAuthProvider, string>;
-const initialOAuthUrlMap: OAuthUrlMap = {
-    google: "",
-    twitter: "",
-    // Farcaster is not included here as it uses a different authentication method
-};
 interface AuthFormContextType {
     step: AuthStep;
     error: string | null;
@@ -20,8 +12,6 @@ interface AuthFormContextType {
     authModalTitle?: string;
     loginMethods: LoginMethod[];
     defaultEmail?: string;
-    oauthUrlMap: OAuthUrlMap;
-    isLoadingOauthUrlMap: boolean;
     baseUrl: string;
     setStep: (step: AuthStep) => void;
     setError: (error: string | null) => void;
@@ -40,7 +30,6 @@ type ContextInitialStateProps = {
 
 type AuthFormProviderProps = {
     setDialogOpen?: (open: boolean, successfulLogin?: boolean) => void;
-    preFetchOAuthUrls: boolean;
     initialState: ContextInitialStateProps;
     children: ReactNode;
 };
@@ -55,51 +44,11 @@ export const useAuthForm = () => {
     return context;
 };
 
-export const AuthFormProvider = ({
-    setDialogOpen,
-    preFetchOAuthUrls,
-    initialState,
-    children,
-}: AuthFormProviderProps) => {
-    const { crossmintAuth } = useCrossmintAuth();
+export const AuthFormProvider = ({ setDialogOpen, initialState, children }: AuthFormProviderProps) => {
     const [step, setStep] = useState<AuthStep>("initial");
     const [error, setError] = useState<string | null>(null);
-    const [oauthUrlMap, setOauthUrlMap] = useState<OAuthUrlMap>(initialOAuthUrlMap);
-    const [isLoadingOauthUrlMap, setIsLoadingOauthUrlMap] = useState(true);
 
     const { loginMethods, baseUrl, appearance, termsOfServiceText, authModalTitle, defaultEmail } = initialState;
-
-    const preFetchAndSetOauthUrl = useCallback(async () => {
-        setIsLoadingOauthUrlMap(true);
-        try {
-            const oauthProviders = loginMethods.filter(
-                (method): method is OAuthProvider => method in initialOAuthUrlMap
-            );
-
-            const oauthPromiseList = oauthProviders.map(async (provider) => {
-                const url = await crossmintAuth?.getOAuthUrl(provider);
-                return { [provider]: url };
-            });
-
-            const oauthUrlMap = Object.assign({}, ...(await Promise.all(oauthPromiseList)));
-            setOauthUrlMap(oauthUrlMap);
-        } catch (error) {
-            setError(
-                error instanceof CrossmintAuthenticationError
-                    ? error.message
-                    : "Unable to load oauth providers. Please try again later."
-            );
-        } finally {
-            setIsLoadingOauthUrlMap(false);
-        }
-    }, [loginMethods, crossmintAuth]);
-
-    useEffect(() => {
-        // Only pre-fetch oauth urls if the user is not logged in
-        if (preFetchOAuthUrls) {
-            preFetchAndSetOauthUrl();
-        }
-    }, [preFetchAndSetOauthUrl, preFetchOAuthUrls]);
 
     const handleToggleDialog = (open: boolean, successfulLogin?: boolean) => {
         setDialogOpen?.(open, successfulLogin);
@@ -118,8 +67,6 @@ export const AuthFormProvider = ({
         authModalTitle,
         loginMethods,
         defaultEmail,
-        oauthUrlMap,
-        isLoadingOauthUrlMap,
         setDialogOpen: handleToggleDialog,
         setError,
         setStep,