+++ apps/payments/vue/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @crossmint/client-sdk-vue-ui-starter
 
+## 1.1.74
+
+### Patch Changes
+
+- @crossmint/client-sdk-base@1.5.14
+- @crossmint/client-sdk-vue-ui@1.1.60
+- @crossmint/client-sdk-smart-wallet@0.3.26
+
 ## 1.1.73
 
 ### Patch Changes