+++ apps/wallets/smart-wallet/next/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @crossmint/client-sdk-smart-wallet-next-starter
 
+## 0.1.149
+
+### Patch Changes
+
+- Updated dependencies [7a12798]
+  - @crossmint/client-sdk-react-ui@2.2.17
+
 ## 0.1.148
 
 ### Patch Changes